% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{flex_dashboard_poster}
\alias{flex_dashboard_poster}
\title{Convert to a poster with fexdashboard layout.}
\usage{
flex_dashboard_poster(fig_width = 6, fig_height = 6, fig_mobile = FALSE,
  dev = "png", self_contained = TRUE, orientation = "rows",
  df_print = "kable", css = NULL, keep_md = FALSE, theme = "default",
  highlight = "default", md_extensions = NULL, ...)
}
\arguments{
\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{fig_mobile}{Create an additional rendering of each R graphics figure
optimized for rendering on mobile devices oriented in portrait mode.
If \code{TRUE}, creates a figure which is 3.75 x 4.80 inches wide;
if \code{FALSE}, create no additional figure for mobile devices;
if a numeric vector of length 2, creates a mobile figure with the
specified width and height.}

\item{dev}{Graphics device to use for figure output (defaults to png)}

\item{self_contained}{Produce a standalone HTML file with no external
dependencies, using data: URIs to incorporate the contents of linked
scripts, stylesheets, images, and videos. Note that even for self contained
documents MathJax is still loaded externally (this is necessary because of
it's size).}

\item{orientation}{Determines whether level 2 headings are treated as
dashboard rows or dashboard columns.}

\item{css}{One or more css files to include}

\item{theme}{Visual theme ("default", "bootstrap", "cerulean", "journal",
"flatly", "readable", "spacelab", "united", "cosmo", "lumen", "paper",
"sandstone", "simplex", or "yeti"). The "cosmo" theme is used when "default"
is specified.}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn",
and "haddock". Pass NULL to prevent syntax highlighting.}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}

\item{...}{Additional arguments passed to \code{\link{flexdashboard::flex_dashboard}}}
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
A flexdashboard poster is an R Markdown document using a flexdashboard
layout to arrange text, figures, tables and other content into a typical
grid format as used for conference posters.
By default, the print version will be rendered with \code{render} when
exiting \code{rmarkdown::render}.
}
\details{
Compared to \code{flexdashboard::flex_dashboard}, it:

\itemize{
  \item uses level 2 headings as dashboard rows by default, because posters are
  usually divided vertically in horizontal sections (e.g. title, body, figures),
  each of which can contain a different number of columns (i.e. level three headings)
  \item prints data frames using \code{\link{knitr::kable}} to show nice simple
  tables when there is data frame output from the code chunks
  \item produces larger figures
  \item never creates mobile rendering of the R graphics
  \item never uses retina figures
 }
}
