% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_classify.R
\name{gp_classify}
\alias{gp_classify}
\title{gp_classify}
\usage{
gp_classify(model, threshold)
}
\arguments{
\item{model}{A glm logistic model}

\item{threshold}{Some numeric threshold}
}
\value{
a vector of classifications
}
\description{
gp_classify
}
