% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_classtable.R
\name{gp_classtable}
\alias{gp_classtable}
\title{Classification Table}
\usage{
gp_classtable(model, threshold, ...)
}
\arguments{
\item{model}{A glm logistic model}

\item{threshold}{Some numeric threshold}

\item{...}{Additional arguments to `table`.}
}
\value{
A table object
}
\description{
Classification Table
}
