% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_rates.R
\name{gp_tpr}
\alias{gp_tpr}
\alias{gp_sensitivity}
\alias{gp_tnr}
\alias{gp_specificity}
\alias{gp_fpr}
\alias{gp_fnr}
\title{gp_tpr}
\usage{
gp_tpr(model, threshold)

gp_sensitivity(model, threshold)

gp_tnr(model, threshold)

gp_specificity(model, threshold)

gp_fpr(model, threshold)

gp_fnr(model, threshold)
}
\arguments{
\item{model}{A glm logistic model}

\item{threshold}{Some numeric threshold}
}
\value{
The true positive rate
}
\description{
gp_tpr
}
