% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnr.R
\name{gp_tnr}
\alias{gp_tnr}
\alias{gp_specificity}
\title{gp_tnr}
\usage{
gp_tnr(model, threshold)

gp_specificity(model, threshold)
}
\arguments{
\item{model}{A glm logistic model}

\item{threshold}{Some numeric threshold}
}
\value{
The true negative rate
}
\description{
gp_tnr
}
