% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpr.R
\name{gp_tpr}
\alias{gp_tpr}
\alias{gp_sensitivity}
\title{gp_tpr}
\usage{
gp_tpr(model, threshold)

gp_sensitivity(model, threshold)
}
\arguments{
\item{model}{A glm logistic model}

\item{threshold}{Some numeric threshold}
}
\value{
The true positive rate
}
\description{
gp_tpr
}
