% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postr.R
\docType{package}
\name{postr-package}
\alias{postr}
\alias{postr-package}
\title{postr: Prepare reproducible posters using R Markdown}
\description{
Provides templates and helper functions to prepare R Markdown based
conference posters and produce output both in html for hosting and png for
printing. Relies on the flexdashboard package to easily define poster layout.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/odeleongt/postr}
  \item Report bugs at \url{https://github.com/odeleongt/postr/issues/}
}

}
\author{
\strong{Maintainer}: Oscar de León \email{odeleon@ces.uvg.edu.gt}

}
