% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{postr_AUC}
\alias{postr_AUC}
\alias{pr_AUC}
\title{Calculate AUC}
\usage{
postr_AUC(model, thresholds = seq(0, 1, by = 0.01))

pr_AUC(model, thresholds = seq(0, 1, by = 0.01))
}
\arguments{
\item{model}{A glm logistic model}

\item{thresholds}{Some numeric thresholds}
}
\value{
AUCm
}
\description{
Calculate AUC
}
