% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificationtable.R
\name{postr_classificationtable}
\alias{postr_classificationtable}
\alias{pr_classificationtable}
\alias{postr_classtable}
\alias{pr_classtable}
\title{Classification Table}
\usage{
postr_classificationtable(model, threshold, ...)

pr_classificationtable(model, threshold, ...)

postr_classtable(model, threshold, ...)

pr_classtable(model, threshold, ...)
}
\arguments{
\item{model}{A glm logistic model}

\item{threshold}{Some numeric threshold}

\item{...}{Additional arguments to `table`.}
}
\value{
A table object
}
\description{
Classification Table
}
