% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postr_classify.R
\name{postr_classify}
\alias{postr_classify}
\alias{pr_classify}
\title{postr_classify}
\usage{
postr_classify(model, threshold)

pr_classify(model, threshold)
}
\arguments{
\item{model}{A glm logistic model}

\item{threshold}{Some numeric threshold}
}
\value{
a vector of classifications
}
\description{
postr_classify
}
