% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leverage.R
\name{postr_leverage}
\alias{postr_leverage}
\alias{pr_leverage}
\title{Return leverage values}
\usage{
postr_leverage(model)

pr_leverage(model)
}
\arguments{
\item{model}{A supported model}
}
\value{
Leverage values
}
\description{
Return leverage values
}
