% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{postr_tpr}
\alias{postr_tpr}
\alias{postr_sensitivity}
\alias{pr_tpr}
\alias{pr_sensitivity}
\alias{postr_tnr}
\alias{postr_specificity}
\alias{pr_tnr}
\alias{pr_specificity}
\alias{postr_fpr}
\alias{pr_fpr}
\alias{postr_fnr}
\alias{pr_fnr}
\title{Classification Rates}
\usage{
postr_tpr(model, threshold)

postr_sensitivity(model, threshold)

pr_tpr(model, threshold)

pr_sensitivity(model, threshold)

postr_tnr(model, threshold)

postr_specificity(model, threshold)

pr_tnr(model, threshold)

pr_specificity(model, threshold)

postr_fpr(model, threshold)

pr_fpr(model, threshold)

postr_fnr(model, threshold)

pr_fnr(model, threshold)
}
\arguments{
\item{model}{A glm logistic model}

\item{threshold}{Some numeric thresholds}
}
\value{
A rate
}
\description{
Returns a classification rate.
}
\details{
`postr_sensitivity` and `postr_specificity` are aliases `postr_tpr` and `postr_tnr`.
}
