% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.r
\name{postr_ROC}
\alias{postr_ROC}
\alias{pr_ROC}
\title{plot ROC curve}
\usage{
postr_ROC(model, thresholds = seq(0, 1, by = 0.01))

pr_ROC(model, thresholds = seq(0, 1, by = 0.01))
}
\arguments{
\item{model}{A glm logistic model}

\item{thresholds}{Some numeric threshold}
}
\value{
ggplot object
}
\description{
plot ROC curve
}
