% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvfplot.R
\name{postr_rvfplot}
\alias{postr_rvfplot}
\alias{pr_rvfplot}
\title{Residual vs Fitted plot}
\usage{
postr_rvfplot(model, smoothedcurve = FALSE)

pr_rvfplot(model, smoothedcurve = FALSE)
}
\arguments{
\item{model}{A model}

\item{smoothedcurve}{Logical; should a smoothed curve be plotted? Default: False}
}
\value{
A ggplot object
}
\description{
Residual vs Fitted plot
}
