% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensplot.R
\name{postr_sensplot}
\alias{postr_sensplot}
\alias{pr_sensplot}
\title{plot Sensitivity (TPR) and Specificity (TNR) curves}
\usage{
postr_sensplot(model, tpr.color = "red", tnr.color = "blue")

pr_sensplot(model, tpr.color = "red", tnr.color = "blue")
}
\arguments{
\item{model}{A glm logistic model}

\item{tpr.color}{Color for sensitivity (TPR) line}

\item{tnr.color}{Color for specificity (TNR) line}
}
\value{
ggplot object
}
\description{
plot Sensitivity (TPR) and Specificity (TNR) curves
}
