% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.families.supported.R
\name{.glm.families.supported}
\alias{.glm.families.supported}
\title{(Internal) function to check GLM/GLMER family support}
\usage{
.glm.families.supported(model, ...)
}
\arguments{
\item{model}{A glm or glmerMod object}

\item{...}{any number of families, e.g. "binomial", "poisson"}
}
\value{
TRUE if family of `model` is in list, otherwise informative
 errors if family of model is not in the list.
}
\description{
(Internal) function to check GLM/GLMER family support
}
