% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{flex_dashboard_poster}
\alias{flex_dashboard_poster}
\title{Convert to a poster with flexdashboard layout.}
\usage{
flex_dashboard_poster(fig_width = 6, fig_height = 6, fig_mobile = FALSE,
  dev = "png", self_contained = TRUE, orientation = "rows",
  df_print = "kable", css = NULL, keep_md = FALSE, theme = "default",
  highlight = "default", md_extensions = NULL, render_print = FALSE, ...)
}
\arguments{
\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{fig_mobile}{Create an additional rendering of each R graphics figure
optimized for rendering on mobile devices oriented in portrait mode.
If \code{TRUE}, creates a figure which is 3.75 x 4.80 inches wide;
if \code{FALSE}, create no additional figure for mobile devices;
if a numeric vector of length 2, creates a mobile figure with the
specified width and height.}

\item{dev}{Graphics device to use for figure output (defaults to png)}

\item{self_contained}{Produce a standalone HTML file with no external
dependencies, using data: URIs to incorporate the contents of linked
scripts, stylesheets, images, and videos. Note that even for self contained
documents MathJax is still loaded externally (this is necessary because of
it's size).}

\item{orientation}{Determines whether level 2 headings are treated as
dashboard rows or dashboard columns.}

\item{df_print}{Method to be used for printing data frames. Valid values
include "default", "kable", "tibble", and "paged". The "default" method uses
\code{print.data.frame}. The "kable" method uses the
\code{\link[knitr:kable]{knitr::kable}} function. The "tibble" method uses
the \pkg{tibble} package to print a summary of the data frame. The "paged"
method creates a paginated HTML table (note that this method is only valid
for formats that produce HTML). In addition
to the named methods you can also pass an arbitrary function to be used
for printing data frames. You can disable the df_print behavior entirely
by setting the option \code{rmarkdown.df_print} to \code{FALSE}.}

\item{css}{One or more css files to include}

\item{keep_md}{Keep the markdown file generated by knitting. Note that
if this is \code{TRUE} then \code{clean_supporting} will always be
\code{FALSE}.}

\item{theme}{Visual theme ("default", "bootstrap", "cerulean", "journal",
"flatly", "readable", "spacelab", "united", "cosmo", "lumen", "paper",
"sandstone", "simplex", or "yeti"). The "cosmo" theme is used when "default"
is specified.}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn",
and "haddock". Pass NULL to prevent syntax highlighting.}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}

\item{render_print}{Whether to render the print version (defaults to FALSE).}

\item{...}{Additional arguments passed to \code{\link[flexdashboard]{flex_dashboard}}}
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
A flexdashboard poster is an R Markdown document using a flexdashboard
layout to arrange text, figures, tables and other content into a typical
grid format as used for conference posters.
The print version will be rendered when knitting if the format option
\code{render_print} is set to \code{TRUE} in the YAML header or the function
call; currently this requires that the root dir be set to the default or a
path relative to the Rmd document location, otherwise it will result in an
error.
}
\details{
Compared to \code{flexdashboard::flex_dashboard}, it:

\itemize{
  \item uses level 2 headings as dashboard rows by default, because posters are
  usually divided vertically in horizontal sections (e.g. title, body, figures),
  each of which can contain a different number of columns (i.e. level three headings)
  \item prints data frames using \code{\link[knitr]{kable}} to show nice simple
  tables when there is data frame output from the code chunks
  \item produces larger figures
  \item never creates mobile rendering of the R graphics
  \item never uses retina figures
 }
}
