% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{postr_classify}
\alias{postr_classify}
\alias{pr_classify}
\title{postr_classify}
\usage{
postr_classify(model, threshold, newdata = NULL)

pr_classify(model, threshold, newdata = NULL)
}
\arguments{
\item{model}{A supported model.}

\item{threshold}{A numeric value in the range [0,1].}

\item{newdata}{An optional dataframe of observations to classify instead of the observations used in building the model.}
}
\value{
A vector of logical classifications.
}
\description{
Given a model and some threshold, estimate the classification for each observation.
}
