% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leverage.R
\name{postr_leverage}
\alias{postr_leverage}
\alias{pr_leverage}
\title{Return leverage values}
\usage{
postr_leverage(model, force = FALSE)

pr_leverage(model, force = FALSE)
}
\arguments{
\item{model}{A supported model}

\item{force}{Force calculation in non-supported models.}
}
\value{
Leverage values
}
\description{
Return leverage values
}
