% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed.R
\name{postr_observed}
\alias{postr_observed}
\title{Return observed values}
\usage{
postr_observed(model)
}
\arguments{
\item{model}{A support model}
}
\value{
Observed values
}
\description{
Return observed values
}
