% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.R
\name{postr_ROC}
\alias{postr_ROC}
\alias{pr_ROC}
\title{plot ROC curve}
\usage{
postr_ROC(model, thresholds = seq(0, 1, by = 0.01), AUC = TRUE)

pr_ROC(model, thresholds = seq(0, 1, by = 0.01), AUC = TRUE)
}
\arguments{
\item{model}{A glm logistic model}

\item{thresholds}{Some numeric threshold}

\item{AUC}{Logical; should AUC be printed on the plot? Default TRUE.}
}
\value{
ggplot object
}
\description{
plot ROC curve
}
